+++ packages/gamut-styles/src/ColorMode.tsx
@@ -35,7 +35,7 @@ export type ColorModeShape = ColorModeConfig[ColorModes];
 export type ColorAlias = keyof ColorModeShape;
 
 export type ColorModeProps = {
-  mode: ColorModes;
+  mode: ColorModes | 'system';
   bg?: Colors;
 };
 
@@ -82,17 +82,19 @@ export function usePrefersDarkMode() {
   const [prefersDarkMode, setPrefersDarkMode] = useState(false);
 
   useEffect(() => {
-    const mq = window.matchMedia('(prefers-color-scheme: dark)');
-
     function onChange(event: MediaQueryListEvent) {
       setPrefersDarkMode(event.matches);
     }
 
-    if (mq && 'addEventListener' in mq) {
-      setPrefersDarkMode(mq.matches);
-      mq.addEventListener('change', onChange);
+    if (window && 'matchMedia' in window) {
+      const mq = window.matchMedia('(prefers-color-scheme: dark)');
 
-      return () => mq.removeEventListener('change', onChange);
+      if (mq && 'addEventListener' in mq) {
+        setPrefersDarkMode(mq.matches);
+        mq.addEventListener('change', onChange);
+
+        return () => mq.removeEventListener('change', onChange);
+      }
     }
   }, []);
 
@@ -112,13 +114,19 @@ export const VariableProvider = styled(
 export const ColorMode = forwardRef<
   HTMLDivElement,
   Omit<ComponentProps<typeof VariableProvider>, 'bg'> & ColorModeProps
->(({ mode, alwaysSetVariables, bg, ...rest }, ref) => {
+>(({ mode: preference, alwaysSetVariables, bg, ...rest }, ref) => {
+  // checks if the user has set 'system' as their color mode preference
+  // then sets the color mode
+  const prefersDarkMode = usePrefersDarkMode();
+  const mode =
+    preference === 'system' ? (prefersDarkMode ? 'dark' : 'light') : preference;
+
   const theme = useTheme();
   const { modes, mode: active, colors } = theme;
   const contextBg = bg ? 'background-current' : undefined;
 
   /** Serialize color variables for the current mode
-   * 1. If all variables are requried add all mode variables to the current context
+   * 1. If all variables are required add all mode variables to the current context
    * 2. If the user has specified a background color - set that color to the current-bg
    * 3. If not
    */